Revert "Simplify the implementation of the eraseInLineFromCursor instruction"
This reverts commit 43d5feee69a197fbeb171ccb3eeff41b4f529143.

The original commit message said:

    There is no need to overwrite things with spaces; since tcell clears the
    whole surface before drawing, it is enough to just truncate the line at the
    current write position.

As it turns out, this is only true if the background color is none; for the case
where where there's a background color set, we do need to draw spaces until the
end of the line. This happens, for example, when using the delta pager in
lazygit: the green and red bars for added and deleted lines didn't extend to the
right edge of the view.

The reason why I made the original commit in the first place is because the
previous implementation had two bugs, and simplifying things the way I did in
43d5feee69 made it unnecessary to fix these. Now that we are reverting the
simplification, we do need to fix those bugs, which I will do in the next three
commits.