Fix SetCursor2D at soft line break
Similar to the code in GoToEndOfLine, we need to move left by one character when
setting the cursor position to a soft line break, so that we end up at the last
whitespace before the line break.

This fixes pressing up-arrow or down-arrow at the end of a line when jumping to
an auto-wrapped line that is shorter.