Only highlight "physical" lines
As opposed to "logical" lines, which means that when a long line is wrapped, we
only highlight the actual view line that the cursor is in, rather than all
wrapped view lines that belong to the original, unwrapped line.

This effectively reverts what was done to fix
https://github.com/jroimartin/gocui/issues/50, but since we don't need (or want)
this feature in lazygit, we don't bother making this configurable.