Search viewLines rather than lines
So far, search was only enabled in views with Wrap=false, so it didn't make a
difference. If Wrap is on, we need to work with viewLines rather than lines,
because that's what the code that highlights the search results expects. Note
that a viewLine index is also passed back to the onSelectItem callback, so
clients need to be able to deal with that.

We don't touch the modelSearchResults code path, as it's a bit unclear how to
support Wrap=true in that case, and we don't need it right now.