Remove Project Gutenberg header to make the file free
The original text of "The Adventures of Tom Sawyer" by Mark Twain is in
the public domain.

However, the eBook version from Project Gutenberg is non-free due to
restrictions on commercial use (specifically clauses 1.E.7, 1.E.8, and
1.E.9 of the Project Gutenberg License).

To make this file free, this commit removes all references to the Project
Gutenberg in accordance with clause 1.C of the Project Gutenberg License.

Full text of the Project Gutenberg License:

    https://www.gutenberg.org/policy/license.html#the-full-project-gutenberg-license-in-legalese-normative

Legal references:

    https://lists.debian.org/debian-legal/2025/03/msg00004.html
    https://lists.debian.org/debian-devel/2025/03/msg00229.html
    https://lists.debian.org/debian-legal/2017/08/msg00001.html

Signed-off-by: Jongmin Kim <jmkim@debian.org>